Crate day_3

Modules

error ğŸ”’

Constants

TASK1_BATTERY_NUM ğŸ”’

TASK2_BATTERY_NUM ğŸ”’

Functions

find_highest_in_slice ğŸ”’

find_highest_joltage ğŸ”’

In order to find valid battery combinations, we use &str slices, with the end point being the length of the string - number of batteries. We keep track of the start and use this as a moving slider. This means that we only need to check the highest digit in each slice (O(n)) in order to get the next valid battery we concatenate each char found to a string and then parse to u64

main ğŸ”’

read_lines ğŸ”’

from Rust By Example - read_lines