Crate day_3

Modules

error 🔒

Functions

find_two_highest 🔒

in order to find the highest pair, we need to find the highest number in all of the line EXCEPT for the last pos - we are finding the A in AB we do this by getting a slice of all but the last character we keep track of the index so we need only check all characters AFTER the A we make a second slice from A position + 1, to the end of the slice finally we concatenate the chars together and parse them to create the numerical battery pair

main 🔒

read_lines 🔒